# server/openapi.yaml
openapi: 3.0.3
info:
  title: ActivityOS API
  version: 0.31.0
  description: |
    Express + Firestore backend for ActivityOS — a **multi-tenant** platform
    for children's activity providers.

    ## v0.31.0 — Cross-provider marketplace (opt-in)

    The parent Browse feed becomes a real marketplace instead of being scoped
    (client-side) to providers a family had already booked.
    - **`GET /api/listings`** (no `?tenantId=`, no `?mine=1`) now returns the
      live/public/unarchived listings of every provider **that has opted in**
      (`settings.marketplaceListed === true`) — **plus**, for a signed-in
      parent, their own providers (anyone they've booked) regardless of opt-in.
    - A provider's **own storefront** (`GET /api/listings?tenantId=`) is
      unchanged — their public listings always show there whether or not
      they're in the marketplace.
    - New tenant setting **`marketplaceListed`** (in the library `settings`
      bag), default **false** — a provider opts in from Setup → Bookings →
      Marketplace. Opt-in is a deliberate choice; nobody is publicly listed
      without it.
    (Frontend: BrowseApp drops the old single-provider scoping and shows the
    server feed as-is.)

    ## v0.30.0 — Account, Data & privacy, parent Accidents

    Clearing the shared/parent mock pages.
    - **`/api/account`** (every role) — the signed-in user's profile. `GET`
      returns `{email, name, phone, marketingConsent, role, tenantId}` (identity
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /api/emails:
    get:
      tags: [emails]
      summary: The send history (operators)
      parameters: [{ name: tenantId, in: query, required: false, description: Platform only, schema: { type: string } }]
      responses:
        "200": { description: Array of sent emails (newest first) }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/emails/recipients:
    get:
      tags: [emails]
      summary: The "all families" audience size + a sample (operators)
      parameters: [{ name: tenantId, in: query, required: false, description: Platform only, schema: { type: string } }]
      responses:
        "200": { description: "{count, sample[]}" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/emails/send:
    post:
      tags: [emails]
      summary: Email families (or dry-run to preview recipients) — operators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, body]
              properties:
                subject: { type: string }
                body: { type: string }
                audience: { type: string, enum: [all, one], default: all }
                to: { type: string, description: "recipient for audience=one" }
                recipients: { type: array, items: { type: string }, description: "explicit list (overrides the all audience)" }
                cc: { type: string, description: "comma-separated; added to the recipient set" }
                bcc: { type: string, description: "comma-separated; added to the recipient set" }
                html: { type: string, description: "pre-rendered designed document, sent as-is" }
                dryRun: { type: boolean, default: false, description: "preview recipients; no send, no record" }
      responses:
        "200": { description: "dry-run: {dryRun:true, recipientCount, sample}" }
        "201": { description: "sent: the recorded email {recipientCount, status:sending, delivered, openedBy, …}; merge fields resolve per recipient" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/emails/audiences:
    get:
      tags: [emails]
      summary: Live CRM segments (all/active/past/waitlisted/enquiries) — operators
      parameters: [{ name: tenantId, in: query, required: false, description: Platform only, schema: { type: string } }]
      responses:
        "200": { description: "[{id, name, desc, count, emails[]}] — membership computed at request time" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/emails/schedule:
    post:
      tags: [emails]
      summary: Queue a send for a future UK wall-clock time (operators)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, body, sendAt]
              properties:
                subject: { type: string }
                body: { type: string }
                html: { type: string }
                audience: { type: string, enum: [all, one], default: all }
                to: { type: string }
                recipients: { type: array, items: { type: string } }
                cc: { type: string }
                bcc: { type: string }
                sendAt: { type: string, description: "YYYY-MM-DDTHH:MM, UK wall clock; must be in the future" }
      responses:
        "201": { description: "the queued send (recipients resolved + frozen now; a 60s sweep fires it at sendAt)" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/emails/scheduled:
    get:
      tags: [emails]
      summary: The scheduled-send queue, soonest first (operators)
      parameters: [{ name: tenantId, in: query, required: false, description: Platform only, schema: { type: string } }]
      responses:
        "200": { description: "[{id, subject, recipientCount, sendAt, status: scheduled|sent|cancelled}] — recipient list not exposed" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/emails/scheduled/{id}:
    delete:
      tags: [emails]
      summary: Cancel a scheduled send that hasn't fired yet (operators)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "{ok:true} — status becomes cancelled" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Already sent }
  /api/emails/messages:
    get:
      tags: [emails]
      summary: The Inbox — received messages, newest first (operators)
      description: An expired snooze wakes its message back into the inbox (unread) as a side effect of this read.
      parameters: [{ name: tenantId, in: query, required: false, description: Platform only, schema: { type: string } }]
      responses:
        "200": { description: "[{id, from, fromEmail?, to?, subject, body, html?, folder, labels[], unread, starred, snoozedUntil, attachments[], at}]" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/emails/messages/{id}:
    patch:
      tags: [emails]
      summary: Star / read / label / move / snooze a message (operators)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folder: { type: string, enum: [inbox, archive, snoozed, spam, trash] }
                unread: { type: boolean }
                starred: { type: boolean }
                labels: { type: array, items: { type: string } }
                snoozedUntil: { type: string, nullable: true, description: "ISO time; setting it moves the message to snoozed" }
      responses:
        "200": { description: The updated message }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [emails]
      summary: Permanently delete a message (operators; the client trashes first)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "{ok:true}" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/messages:
    post:
      tags: [messages]
      summary: Send a message (creates the thread if needed) — booking-gated
      description: >
        Operator body `{parentEmail, parentName?, body}` (the family must be a
        customer, else 400). Parent body `{tenantId, body}` (must have booked
        that provider, else 403).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                parentEmail: { type: string, description: "operator only" }
                parentName: { type: string, description: "operator only" }
                tenantId: { type: string, description: "parent only" }
      responses:
        "201": { description: "The created message (+ threadId)" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token (see the Authentication section above).
...
  responses:
    BadRequest:
      description: Validation failed (`error` is a message or zod issue list)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: The account's role/tenant doesn't allow this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or outside the caller's tenant scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Human-readable message, or an array of zod issues on 400.
          oneOf:
            - type: string
            - type: array
              items: { type: object }
...